requirements.txt
# core
langchain
langchain-community
langchain-text-splitters
langchain-core

# llm
langchain-groq

# embeddings
sentence-transformers
langchain-huggingface

# vector DB
langchain-chroma
chromadb

# google drive
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2

# app
streamlit
python-dotenv